updated directory structure
