+++ gui/include/ConfigureDialog.h
