+++ libmatrix/MatrixManager.h
